Move functions to utils; Create new fuzzer
